refactor: separate order list and detail API routes
